+++ Manifest.toml
@@ -2,7 +2,7 @@
 
 julia_version = "1.12.5"
 manifest_format = "2.0"
-project_hash = "405c68880b86e28d09ee63d15adbf38ae3ae77c5"
+project_hash = "966399a65a106adeb69d0092ad1fcaa13e09960d"
 
 [[deps.AbstractFFTs]]
 deps = ["LinearAlgebra"]
@@ -1830,12 +1830,6 @@ deps = ["Distributed", "Mmap", "Random", "Serialization"]
 uuid = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
 version = "1.11.0"
 
-[[deps.SharedHydrology]]
-deps = ["StaticArrays"]
-path = "SharedHydrology"
-uuid = "e4c7f7b8-5a3d-4e6f-9c1a-2b8d3f4e5a6c"
-version = "0.1.0-dev"
-
 [[deps.Showoff]]
 deps = ["Dates", "Grisu"]
 git-tree-sha1 = "91eddf657aca81df9ae6ceb20b959ae5653ad1de"
@@ -2181,7 +2175,7 @@ uuid = "e3aaa7dc-3e4b-44e0-be63-ffb868ccd7c1"
 version = "0.1.3"
 
 [[deps.Wflow]]
-deps = ["Accessors", "BasicModelInterface", "CFTime", "CompositionsBase", "Dates", "DelimitedFiles", "EnumX", "FillArrays", "Glob", "Graphs", "LoggingExtras", "NCDatasets", "OrderedCollections", "Polyester", "ProgressLogging", "PropertyDicts", "SharedHydrology", "StaticArrays", "Statistics", "TOML", "TerminalLoggers"]
+deps = ["Accessors", "BasicModelInterface", "CFTime", "CompositionsBase", "Dates", "DelimitedFiles", "EnumX", "FillArrays", "Glob", "Graphs", "LoggingExtras", "NCDatasets", "OrderedCollections", "Polyester", "ProgressLogging", "PropertyDicts", "StaticArrays", "Statistics", "TOML", "TerminalLoggers", "WflowSoil"]
 path = "Wflow"
 uuid = "d48b7d99-76e7-47ae-b1d5-ff0c1cf9a818"
 version = "1.1.0-dev"
@@ -2191,6 +2185,12 @@ deps = ["JSON3", "Logging", "StructTypes", "Wflow", "ZMQ"]
 path = "server"
 uuid = "6f3aa2dc-f527-41f4-85fc-b3cbabe74cf6"
 version = "0.1.0"
+
+[[deps.WflowSoil]]
+deps = ["FillArrays", "StaticArrays"]
+path = "WflowSoil"
+uuid = "e4c7f7b8-5a3d-4e6f-9c1a-2b8d3f4e5a6c"
+version = "0.1.0-dev"
 
 [[deps.WoodburyMatrices]]
 deps = ["LinearAlgebra", "SparseArrays"]
