--- SharedHydrology/src/SharedHydrology.jl
@@ -1,11 +1,0 @@
-module SharedHydrology
-
-using StaticArrays: SVector
-
-const MISSING_VALUE = Float64(NaN)
-
-include("vegetation.jl")
-include("soil/soil_structs.jl")
-include("soil/soil_process.jl")
-
-end # SharedHydrology
