+++ Wflow/Project.toml
@@ -23,7 +23,7 @@ OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 ProgressLogging = "33c8b6b6-d38a-422a-b730-caa89a2f386c"
 PropertyDicts = "f8a19df8-e894-5f55-a973-672c1158cbca"
-SharedHydrology = "e4c7f7b8-5a3d-4e6f-9c1a-2b8d3f4e5a6c"
+WflowSoil = "e4c7f7b8-5a3d-4e6f-9c1a-2b8d3f4e5a6c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
@@ -47,8 +47,8 @@ Polyester = "0.7.10"
 ProgressLogging = "0.1"
 PropertyDicts = "0.2.1"
 StaticArrays = "0.12,1.0"
-SharedHydrology = "0.1"
 Statistics = "1"
 TOML = "1"
+WflowSoil = "0.1"
 TerminalLoggers = "0.1.5"
 julia = "1.10"
