+++ Wflow/src/soil/soil.jl
@@ -1,5 +1,5 @@
 "Initialize SBM soil model variables"
-function SharedHydrology.SbmSoilVariables(n::Int, parameters::SbmSoilParameters)
+function WflowSoil.SbmSoilVariables(n::Int, parameters::SbmSoilParameters)
     (;
         soil_thickness,
         maximum_number_of_layers,
@@ -129,7 +129,7 @@ function sbm_kv_profiles(
 end
 
 "Initialize SBM soil model parameters"
-function SharedHydrology.SbmSoilParameters(
+function WflowSoil.SbmSoilParameters(
     dataset::NCDataset,
     config::Config,
     vegetation_parameter_set::VegetationParameters,
@@ -430,7 +430,7 @@ function SharedHydrology.SbmSoilParameters(
 end
 
 "Initialize SBM soil model"
-function SharedHydrology.SbmSoilModel(
+function WflowSoil.SbmSoilModel(
     dataset::NCDataset,
     config::Config,
     vegetation_parameter_set::VegetationParameters,
@@ -487,6 +487,21 @@ function update_bc_soil_model!(
     return nothing
 end
 
+"""
+    accumulate_open_water_evapotranspiration!(soil_model, runoff, demand)
+
+Accumulate open water evaporation from `runoff` and paddy evaporation from `demand` into
+the total actual evapotranspiration of `soil_model`. Call after `update_soil_water_flow!`.
+"""
+function accumulate_open_water_evapotranspiration!(soil_model::SbmSoilModel, runoff, demand)
+    f = soil_model.variables.fluxes
+    @. f.actual_evapotranspiration +=
+        runoff.variables.actual_open_water_evaporation_river +
+        runoff.variables.actual_open_water_evaporation_land
+    f.actual_evapotranspiration .+= get_evaporation(demand.paddy)
+    return nothing
+end
+
 "Update soil temperature of the SBM soil model for a single timestep"
 function soil_temperature!(
     soil_model::SbmSoilModel,
@@ -501,539 +516,6 @@ function soil_temperature!(
 end
 
 soil_temperature!(::SbmSoilModel, ::NoSnowModel, ::Vector{Float64}) = nothing
-
-"Update total available water in the unsaturated zone of the SBM soil model for a single timestep"
-function unsaturated_store_depth!(soil_model::SbmSoilModel)
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    p = soil_model.parameters
-    for i in eachindex(s.unsaturated_layer_depth)
-        d.unsaturated_store_depth[i] =
-            sum(@view s.unsaturated_layer_depth[i][1:p.number_of_layers[i]])
-    end
-    return nothing
-end
-
-"Update the infiltration reduction factor of the SBM soil model for a single timestep"
-function infiltration_reduction_factor!(
-    soil_model::SbmSoilModel;
-    modelsnow = false,
-    soil_infiltration_reduction = false,
-)
-    s = soil_model.variables.states
-    im = soil_model.variables.intermediates
-    p = soil_model.parameters
-
-    n = length(s.soil_surface_temperature)
-    threaded_foreach(1:n; basesize = 1000) do i
-        im.f_infiltration_reduction[i] = infiltration_reduction_factor(
-            s.soil_surface_temperature[i],
-            p.cf_soil[i];
-            modelsnow,
-            soil_infiltration_reduction,
-        )
-    end
-    return nothing
-end
-
-"""
-    infiltration!(soil_model::SbmSoilModel, dt::Float64)
-
-Update the infiltration rate `infiltration` and infiltration excess water rate
-`infiltration_excess` of the SBM soil model for a single timestep.
-"""
-function infiltration!(soil_model::SbmSoilModel, dt::Float64)
-    d = soil_model.variables.diagnostic
-    im = soil_model.variables.intermediates
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-    (; water_flux_surface) = soil_model.boundary_conditions
-
-    n = length(f.infiltration)
-    threaded_foreach(1:n; basesize = 1000) do i
-        f.infiltration[i], f.infiltration_excess[i] = infiltration(
-            water_flux_surface[i],
-            p.compacted_soil_area_fraction[i],
-            p.infiltration_capacity_soil[i],
-            p.infiltration_capacity_compacted_soil[i],
-            d.unsaturated_store_capacity[i],
-            im.f_infiltration_reduction[i],
-            dt,
-        )
-    end
-    return nothing
-end
-
-"""
-    unsaturated_zone_flow!(soil_model::SbmSoilModel, dt::Float64)
-
-Update unsaturated storage `unsaturated_layer_depth` and the `transfer` of water from the unsaturated
-to the saturated store of the SBM soil model for a single timestep, based on the Brooks-Corey
-approach.
-"""
-function unsaturated_zone_flow!(soil_model::SbmSoilModel, dt::Float64)
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-
-    n = length(f.transfer)
-    threaded_foreach(1:n; basesize = 250) do i
-        if d.n_unsatlayers[i] > 0
-            # Brooks-Corey approach
-            z = cumsum(d.unsaturated_layer_thickness[i])
-            flow_rate = 0.0
-            for m in 1:d.n_unsatlayers[i]
-                l_sat = d.unsaturated_layer_thickness[i][m] * (p.theta_s[i] - p.theta_r[i])
-                kv_z = hydraulic_conductivity_at_depth(
-                    p.kv_profile,
-                    p.vertical_hydraulic_conductivity_factor,
-                    z[m],
-                    i,
-                    m,
-                )
-                unsaturated_layer_depth = if m == 1
-                    s.unsaturated_layer_depth[i][m] + f.infiltration[i] * dt
-                else
-                    s.unsaturated_layer_depth[i][m] + flow_rate * dt
-                end
-                unsaturated_layer_depth, flow_rate = unsatzone_flow_layer(
-                    unsaturated_layer_depth,
-                    kv_z,
-                    l_sat,
-                    p.brooks_corey_exponent[i][m],
-                    dt,
-                )
-                s.unsaturated_layer_depth[i] =
-                    setindex(s.unsaturated_layer_depth[i], unsaturated_layer_depth, m)
-            end
-            f.transfer[i] = flow_rate
-        else
-            f.transfer[i] = 0.0
-        end
-    end
-    return nothing
-end
-
-"""
-    soil_evaporation!(soil_model::SbmSoilModel)
-
-Update soil evaporation from the saturated store `soil_evaporation_saturated_zone` and the total soil
-evaporation from the unsaturated and saturated store `soil_evaporation` of the SBM soil model for a
-single timestep. Also unsaturated storage `unsaturated_layer_depth` and the saturated store
-`saturated_water_depth` are updated.
-"""
-function soil_evaporation!(soil_model::SbmSoilModel, dt::Float64)
-    (; potential_soilevaporation) = soil_model.boundary_conditions
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-
-    n = length(potential_soilevaporation)
-    threaded_foreach(1:n; basesize = 1000) do i
-        potsoilevap = potential_soilevaporation[i]
-        # First calculate the evaporation of unsaturated storage into the
-        # atmosphere from the upper layer.
-        soilevapunsat = soil_evaporation_unsaturated_store(
-            potsoilevap,
-            s.unsaturated_layer_depth[i][1],
-            d.unsaturated_layer_thickness[i][1],
-            d.n_unsatlayers[i],
-            d.water_table_depth[i],
-            p.theta_s[i] - p.theta_r[i],
-        )
-        # Ensure that the unsaturated evaporation rate does not exceed the
-        # available unsaturated moisture
-        soilevapunsat = min(soilevapunsat, s.unsaturated_layer_depth[i][1] / dt)
-        # Update the additional atmospheric demand
-        potsoilevap -= soilevapunsat
-        s.unsaturated_layer_depth[i] = setindex(
-            s.unsaturated_layer_depth[i],
-            s.unsaturated_layer_depth[i][1] - soilevapunsat * dt,
-            1,
-        )
-        theta_drainable = lower_bound_drainable_porosity(p.theta_s[i], p.theta_fc[i])
-        soil_evaporation_saturated_zone = soil_evaporation_saturated_store(
-            potsoilevap,
-            d.n_unsatlayers[i],
-            p.actual_layer_thickness[i][1],
-            d.water_table_depth[i],
-            theta_drainable,
-            dt,
-        )
-        f.soil_evaporation_saturated_zone[i] = soil_evaporation_saturated_zone
-        f.soil_evaporation[i] = soilevapunsat + soil_evaporation_saturated_zone
-        d.drainable_water_depth[i] -= soil_evaporation_saturated_zone * dt
-    end
-    return nothing
-end
-
-"""
-    transpiration!(soil_model::SbmSoilModel, dt::Float64)
-
-Update total `transpiration`, transpiration from the unsaturated store `actual_evaporation_unsaturated_store` and
-saturated store `actual_evaporation_saturated_zone` of the SBM soil model for a single timestep. Also unsaturated
-storage `unsaturated_layer_depth` and the saturated store `saturated_water_depth` are updated.
-"""
-function transpiration!(soil_model::SbmSoilModel, dt::Float64)
-    (; potential_transpiration) = soil_model.boundary_conditions
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    im = soil_model.variables.intermediates
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-
-    rooting_depth = get_rootingdepth(soil_model)
-    n = length(rooting_depth)
-
-    threaded_foreach(1:n; basesize = 250) do i
-        im.h3[i] = feddes_h3(p.h3_high[i], p.h3_low[i], potential_transpiration[i])
-
-        # compute sum of root fraction in unsaturated soil layers and adapt root fraction
-        # lowest unsaturated soil layer if water table depth intersects the unsaturated root
-        # zone
-        sum_rootfraction_unsat = 0.0
-        rootfraction_unsat_lowest = 0.0
-        for k in 1:d.n_unsatlayers[i]
-            # the root fraction is valid for the root length in a soil layer, if zi decreases
-            # the root length the root fraction needs to be adapted
-            if k == d.n_unsatlayers[i] && d.water_table_depth[i] < rooting_depth[i]
-                rootlength = min(
-                    p.actual_layer_thickness[i][k],
-                    rooting_depth[i] - p.cumulative_layer_depth[i][k],
-                )
-                rootfraction_unsat =
-                    p.rootfraction[i][k] *
-                    (d.unsaturated_layer_thickness[i][k] / rootlength)
-            else
-                rootfraction_unsat = p.rootfraction[i][k]
-            end
-            sum_rootfraction_unsat += rootfraction_unsat
-
-            # rootfraction lowest unsaturated layer
-            rootfraction_unsat_lowest = rootfraction_unsat
-        end
-
-        actevapustore = 0.0
-        for k in 1:d.n_unsatlayers[i]
-            # scale rootfraction soil layer unsaturated zone based on sum of rootfraction in
-            # unsaturated zone
-            if k < d.n_unsatlayers[i]
-                rootfraction_unsat = p.rootfraction[i][k]
-            else
-                rootfraction_unsat = rootfraction_unsat_lowest
-            end
-            rootfraction_unsat_scaled =
-                rooting_depth[i] > 0.0 ?
-                max((1.0 / sum_rootfraction_unsat), 1.0) * rootfraction_unsat : 0.0
-            volumetric_water_content = max(
-                s.unsaturated_layer_depth[i][k] / d.unsaturated_layer_thickness[i][k],
-                1e-7,
-            )
-            head = head_brooks_corey(
-                volumetric_water_content,
-                p.theta_s[i],
-                p.theta_r[i],
-                p.brooks_corey_exponent[i][k],
-                p.air_entry_pressure[i],
-            )
-            alpha = rwu_reduction_feddes(
-                head,
-                p.h1[i],
-                p.h2[i],
-                im.h3[i],
-                p.h4[i],
-                p.alpha_h1[i],
-            )
-            availcap = min(
-                1.0,
-                max(
-                    0.0,
-                    (rooting_depth[i] - p.cumulative_layer_depth[i][k]) /
-                    d.unsaturated_layer_thickness[i][k],
-                ),
-            )
-            maxextr = s.unsaturated_layer_depth[i][k] * availcap / dt
-            actevapustore_layer =
-                min(alpha * rootfraction_unsat_scaled * potential_transpiration[i], maxextr)
-            unsaturated_layer_depth =
-                s.unsaturated_layer_depth[i][k] - actevapustore_layer * dt
-            actevapustore += actevapustore_layer
-            s.unsaturated_layer_depth[i] =
-                setindex(s.unsaturated_layer_depth[i], unsaturated_layer_depth, k)
-        end
-
-        # transpiration from saturated store
-        wetroots = scurve(
-            d.water_table_depth[i],
-            rooting_depth[i],
-            1.0,
-            p.wet_root_distribution_parameter[i],
-        )
-        alpha = rwu_reduction_feddes(
-            Float64(0.0),
-            p.h1[i],
-            p.h2[i],
-            im.h3[i],
-            p.h4[i],
-            p.alpha_h1[i],
-        )
-        restpottrans = potential_transpiration[i] - actevapustore
-        actual_evaporation_saturated_zone =
-            min(restpottrans * wetroots * alpha, d.drainable_water_depth[i] / dt)
-
-        f.actual_evaporation_unsaturated_store[i] = actevapustore
-        f.actual_evaporation_saturated_zone[i] = actual_evaporation_saturated_zone
-        d.drainable_water_depth[i] -= actual_evaporation_saturated_zone * dt
-        f.transpiration[i] = actevapustore + actual_evaporation_saturated_zone
-    end
-    return nothing
-end
-
-"""
-    actual_infiltration!(soil_model::SbmSoilModel)
-
-Update the actual infiltration rate `actual_infiltration` of the SBM soil model for a single timestep.
-
-A soil water balance check is performed. Unsaturated storage that exceeds the maximum
-storage per unsaturated soil layer is transferred to the layer above (or surface), from the
-bottom to the top unsaturated soil layer. The resulting excess water `ustoredepth_excess` is
-subtracted from the infiltration rate `infiltration`.
-"""
-function actual_infiltration!(soil_model::SbmSoilModel, dt::Float64)
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-
-    n = length(f.actual_infiltration)
-    threaded_foreach(1:n; basesize = 1000) do i
-        # check soil moisture balance per layer
-        ustoredepth_excess = 0.0
-        for k in d.n_unsatlayers[i]:-1:1
-            ustoredepth_excess = max(
-                0.0,
-                s.unsaturated_layer_depth[i][k] -
-                d.unsaturated_layer_thickness[i][k] * (p.theta_s[i] - p.theta_r[i]),
-            )
-            s.unsaturated_layer_depth[i] = setindex(
-                s.unsaturated_layer_depth[i],
-                s.unsaturated_layer_depth[i][k] - ustoredepth_excess,
-                k,
-            )
-            if k > 1
-                s.unsaturated_layer_depth[i] = setindex(
-                    s.unsaturated_layer_depth[i],
-                    s.unsaturated_layer_depth[i][k - 1] + ustoredepth_excess,
-                    k - 1,
-                )
-            end
-        end
-        f.actual_infiltration[i] = f.infiltration[i] - ustoredepth_excess / dt
-    end
-    return nothing
-end
-
-"""
-    actual_infiltration_soil_path!(soil_model::SbmSoilModel)
-
-Update the actual infiltration rate for soil `actual_infiltration_soil` and paved area `actual_infiltration_compacted_soil`
-of the SBM soil model for a single timestep.
-"""
-function actual_infiltration_soil_path!(soil_model::SbmSoilModel)
-    im = soil_model.variables.intermediates
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-    (; water_flux_surface) = soil_model.boundary_conditions
-
-    n = length(water_flux_surface)
-    threaded_foreach(1:n; basesize = 1000) do i
-        f.actual_infiltration_soil[i], f.actual_infiltration_compacted_soil[i] =
-            actual_infiltration_soil_path(
-                water_flux_surface[i],
-                f.actual_infiltration[i],
-                p.compacted_soil_area_fraction[i],
-                p.infiltration_capacity_soil[i],
-                p.infiltration_capacity_compacted_soil[i],
-                im.f_infiltration_reduction[i],
-            )
-    end
-    return nothing
-end
-
-"""
-    capillary_flux!(soil_model::SbmSoilModel)
-
-Update the capillary flux `actual_capillary_flux` of the SBM soil model for a single timestep.
-"""
-function capillary_flux!(soil_model::SbmSoilModel, dt::Float64)
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-    rooting_depth = get_rootingdepth(soil_model)
-
-    n = length(rooting_depth)
-    threaded_foreach(1:n; basesize = 1000) do i
-        if d.n_unsatlayers[i] > 0
-            ksat = hydraulic_conductivity_at_depth(
-                p.kv_profile,
-                p.vertical_hydraulic_conductivity_factor,
-                d.water_table_depth[i],
-                i,
-                d.n_unsatlayers[i],
-            )
-            maxcapflux = max(
-                0.0,
-                min(
-                    ksat,
-                    f.actual_evaporation_unsaturated_store[i],
-                    d.unsaturated_store_capacity[i] / dt,
-                    d.drainable_water_depth[i] / dt,
-                ),
-            )
-
-            capflux = if d.water_table_depth[i] > rooting_depth[i]
-                maxcapflux * pow(
-                    1.0 - min(d.water_table_depth[i], p.cap_hmax[i]) / (p.cap_hmax[i]),
-                    p.cap_n[i],
-                )
-            else
-                0.0
-            end
-            netcapflux = capflux
-            actual_capillary_flux = 0.0
-            for k in d.n_unsatlayers[i]:-1:1
-                toadd = min(
-                    netcapflux,
-                    max(
-                        (
-                            d.unsaturated_layer_thickness[i][k] *
-                            (p.theta_s[i] - p.theta_r[i]) -
-                            s.unsaturated_layer_depth[i][k]
-                        ) / dt,
-                        0.0,
-                    ),
-                )
-                s.unsaturated_layer_depth[i] = setindex(
-                    s.unsaturated_layer_depth[i],
-                    s.unsaturated_layer_depth[i][k] + toadd * dt,
-                    k,
-                )
-                netcapflux -= toadd
-                actual_capillary_flux += toadd
-            end
-            f.actual_capillary_flux[i] = actual_capillary_flux
-        else
-            f.actual_capillary_flux[i] = 0.0
-        end
-    end
-    return nothing
-end
-
-"""
-    leakage!(soil_model::SbmSoilModel, dt::Float64)
-
-Update the actual leakage rate `actual_leakage` of the SBM soil model for a single timestep.
-"""
-function leakage!(soil_model::SbmSoilModel, dt::Float64)
-    d = soil_model.variables.diagnostic
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-
-    n = length(f.actual_leakage)
-    threaded_foreach(1:n; basesize = 1000) do i
-        deepksat = hydraulic_conductivity_at_depth(
-            p.kv_profile,
-            p.vertical_hydraulic_conductivity_factor,
-            p.soil_thickness[i],
-            i,
-            p.number_of_layers[i],
-        )
-
-        deeptransfer = min(d.drainable_water_depth[i] / dt, deepksat)
-        f.actual_leakage[i] = max(0.0, min(p.maximum_leakage[i], deeptransfer))
-    end
-    return nothing
-end
-
-"""
-    update_soil_water_flow!(
-        soil_model::SbmSoilModel,
-        atmospheric_forcing::AtmosphericForcing,
-        external_models::NamedTuple,
-        config::Config,
-        dt::Float64,
-    )
-
-Update the SBM soil model (infiltration, unsaturated zone flow, soil evaporation and
-transpiration, capillary flux and leakage) for a single timestep.
-"""
-function update_soil_water_flow!(
-    soil_model::SbmSoilModel,
-    atmospheric_forcing::AtmosphericForcing,
-    external_models::NamedTuple,
-    config::Config,
-    dt::Float64,
-)
-    (; snow, runoff, demand) = external_models
-    (; temperature) = atmospheric_forcing
-    (; water_flux_surface) = soil_model.boundary_conditions
-    s = soil_model.variables.states
-    d = soil_model.variables.diagnostic
-    f = soil_model.variables.fluxes
-    p = soil_model.parameters
-
-    # mainly required for external state changes (e.g. through BMI)
-    update_diagnostic_vars!(soil_model)
-    # infiltration
-    soil_temperature!(soil_model, snow, temperature)
-    infiltration_reduction_factor!(
-        soil_model;
-        modelsnow = config.model.snow__flag,
-        soil_infiltration_reduction = config.model.soil_infiltration_reduction__flag,
-    )
-    infiltration!(soil_model, dt)
-    # unsaturated zone flow
-    unsaturated_zone_flow!(soil_model, dt)
-    # soil evaporation and transpiration
-    soil_evaporation!(soil_model, dt)
-    transpiration!(soil_model, dt)
-    # actual infiltration and excess water
-    actual_infiltration!(soil_model, dt)
-    @. f.saturation_excess_water =
-        (water_flux_surface - f.actual_infiltration) - f.infiltration_excess
-    actual_infiltration_soil_path!(soil_model)
-    @. f.excess_water_soil = max(
-        water_flux_surface * (1.0 - p.compacted_soil_area_fraction) -
-        f.actual_infiltration_soil,
-        0.0,
-    )
-    @. f.excess_water_compacted_soil = max(
-        water_flux_surface * p.compacted_soil_area_fraction -
-        f.actual_infiltration_compacted_soil,
-        0.0,
-    )
-    # recompute the unsaturated store and unsaturated_store_capacity (for capillary flux)
-    unsaturated_store_depth!(soil_model)
-    @. d.unsaturated_store_capacity =
-        p.soil_water_capacity - s.saturated_water_depth - d.unsaturated_store_depth
-    # capillary flux and leakage
-    capillary_flux!(soil_model, dt)
-    leakage!(soil_model, dt)
-    # recharge rate to the saturated store
-    @. f.recharge = (
-        f.transfer - f.actual_capillary_flux - f.actual_leakage -
-        f.actual_evaporation_saturated_zone - f.soil_evaporation_saturated_zone
-    )
-    # total actual evapotranspiration
-    f.actual_evapotranspiration .=
-        f.soil_evaporation .+ f.transpiration .+
-        runoff.variables.actual_open_water_evaporation_river .+
-        runoff.variables.actual_open_water_evaporation_land .+ get_evaporation(demand.paddy)
-    return nothing
-end
 
 function update_ustorelayerdepth!(soil, zi_prev, water_table_depth, i)
     s = soil.variables.states
@@ -1218,51 +700,3 @@ function update_soil_water_storage!(
     @. f.net_runoff = f.runoff - actual_open_water_evaporation_land
     return nothing
 end
-
-"""
-    update_diagnostic_vars!(soil_model::SbmSoilModel)
-
-Update diagnostic variables of `SbmSoilModel` that are critical for subsequent computations
-and depend on state variables `saturated_water_depth` and `unsaturated_layer_depth`.
-"""
-function update_diagnostic_vars!(soil_model::SbmSoilModel)
-    (; saturated_water_depth) = soil_model.variables.states
-    (;
-        water_table_depth,
-        drainable_water_depth,
-        unsaturated_layer_thickness,
-        unsaturated_store_capacity,
-        unsaturated_store_depth,
-        total_soil_water_storage,
-        n_unsatlayers,
-    ) = soil_model.variables.diagnostic
-    (;
-        soil_thickness,
-        theta_s,
-        theta_r,
-        theta_fc,
-        soil_water_capacity,
-        cumulative_layer_depth,
-        actual_layer_thickness,
-    ) = soil_model.parameters
-
-    unsaturated_store_depth!(soil_model)
-    @. water_table_depth =
-        max(0.0, soil_thickness - saturated_water_depth / (theta_s - theta_r))
-    @. drainable_water_depth =
-        (soil_thickness - water_table_depth) *
-        lower_bound_drainable_porosity(theta_s, theta_fc)
-    @. unsaturated_store_capacity =
-        soil_water_capacity - saturated_water_depth - unsaturated_store_depth
-    @. unsaturated_layer_thickness = set_layerthickness(
-        water_table_depth,
-        cumulative_layer_depth,
-        actual_layer_thickness,
-    )
-    @. n_unsatlayers = number_of_active_layers(unsaturated_layer_thickness)
-    @. total_soil_water_storage = saturated_water_depth + unsaturated_store_depth
-end
-
-# wrapper method
-get_rootingdepth(soil_model::SbmSoilModel) =
-    soil_model.parameters.vegetation_parameter_set.rooting_depth
