+++ Wflow/src/utils.jl
@@ -12,23 +12,6 @@ const PCR_DIR = [
     CartesianIndex(1, 1),  # 9
 ]
 
-"""
-    scurve(x, a, b, c)
-
-Sigmoid "S"-shaped curve.
-
-# Arguments
-- `x::Real`: input
-- `a::Real`: determines the center level
-- `b::Real`: determines the amplitude of the curve (range: (0, b⁻¹))
-- `c::Real`: determines the steepness or "stepwiseness" of the curve.
-             The higher c the sharper the function. A negative c reverses the function.
-"""
-function scurve(x::Real, a::Real, b::Real, c::Real)::Real
-    s = inv(b + exp(-c * (x - a)))
-    return s
-end
-
 function to_enumx(T, i::Int)
     options = instances(T)
     n_options = length(options)
@@ -380,34 +363,6 @@ function ncread(
 end
 
 """
-    set_layerthickness(reference_depth::Real, cum_depth::SVector, thickness::SVector)
-
-Calculate actual soil thickness of layers based on a reference depth (e.g. soil depth or
-water table depth) `reference_depth`, a SVector `cum_depth` with cumulative soil depth starting
-at soil surface (0), and a SVector `thickness` with thickness per soil layer.
-"""
-function set_layerthickness(
-    reference_depth::Real,
-    cum_depth::SVector,
-    thickness::SVector{N, Float64},
-)::SVector{N, Float64} where {N}
-    thicknesslayers = thickness .* MISSING_VALUE
-    for i in 1:length(thicknesslayers)
-        if reference_depth > cum_depth[i + 1]
-            thicknesslayers = setindex(thicknesslayers, thickness[i], i)
-        elseif reference_depth - cum_depth[i] > 0.0
-            thicknesslayers = setindex(thicknesslayers, reference_depth - cum_depth[i], i)
-        end
-    end
-    return thicknesslayers
-end
-
-function number_of_active_layers(thickness::SVector)::Int
-    number_of_layers = length(thickness) - sum(isnan.(thickness))
-    return number_of_layers
-end
-
-"""
     get_flow_length(ldd, x_length, y_length)
 
 Return the flow length for a non square grid. Input `ldd` (drainage network), `x_length`
@@ -463,10 +418,6 @@ function get_surface_width(
     surface_width = river_location ? land_area / flow_length : flow_width
     return surface_width
 end
-
-# 2.5x faster power method
-"Faster method for exponentiation"
-pow(x::Real, y::Real)::Real = exp(y * log(x))
 
 function sum_at(A::AbstractVector{T}, inds::AbstractVector{Int})::T where {T}
     mapreduce(i -> A[i], +, inds; init = zero(T))
@@ -674,117 +625,6 @@ function julian_day(time::TimeType)::Int
     # for all years February 28 is day 59 and March 1 is day 60.
     day = dayofyear(time) - (isleapyear(time) && dayofyear(time) > 60)
     return day
-end
-
-"Partition indices with at least size `basesize`"
-function _partition(xs::Integer, basesize::Integer)
-    n = Int(max(1, xs ÷ basesize))
-    return (Int(1 + ((i - 1) * xs) ÷ n):Int((i * xs) ÷ n) for i in 1:n)
-end
-
-"""
-    threaded_foreach(f, x::AbstractArray; basesize::Integer)
-
-Run function `f` in parallel by spawning tasks (nthreads <= 8), each task iterates over a
-chunk of size `basesize`. For nthreads > 8 run function `f` in parallel with
-`Polyester@batch` with `minbatch` equal to `basesize`.
-"""
-function threaded_foreach(f::Function, x::AbstractArray; basesize::Integer)::Nothing
-    if Threads.nthreads() <= 8
-        len = length(x)
-        partitions = _partition(len, basesize)
-        if length(partitions) > 1 && Threads.nthreads() > 1
-            @sync for p in partitions
-                Threads.@spawn begin
-                    for i in eachindex(p)
-                        f(@inbounds p[i])
-                    end
-                end
-            end
-        else
-            for i in eachindex(x)
-                f(@inbounds x[i])
-            end
-        end
-    else
-        @batch per = thread minbatch = basesize for i in eachindex(x)
-            f(@inbounds x[i])
-        end
-    end
-    return nothing
-end
-
-"""
-    hydraulic_conductivity_at_depth(p::KvExponential, vertical_hydraulic_conductivity_factor, z, i, n)
-    hydraulic_conductivity_at_depth(p::KvExponentialConstant, vertical_hydraulic_conductivity_factor, z, i, n)
-    hydraulic_conductivity_at_depth(p::KvLayered, vertical_hydraulic_conductivity_factor, z, i, n)
-    hydraulic_conductivity_at_depth(p::KvLayeredExponential, vertical_hydraulic_conductivity_factor, z, i, n)
-
-Return vertical hydraulic conductivity `kv_z` at depth `z` for index `i` using multiplication
-factor `kv_frac` at soil layer `n` and vertical hydraulic conductivity profile `p`.
-"""
-function hydraulic_conductivity_at_depth(
-    p::KvExponential,
-    vertical_hydraulic_conductivity_factor,
-    z,
-    i,
-    n,
-)
-    kv_z =
-        vertical_hydraulic_conductivity_factor[i][n] *
-        p.kv_0[i] *
-        exp(-p.hydraulic_conductivity_scale_parameter[i] * z)
-    return kv_z
-end
-
-function hydraulic_conductivity_at_depth(
-    p::KvExponentialConstant,
-    vertical_hydraulic_conductivity_factor,
-    z,
-    i,
-    n,
-)
-    (; kv_0, hydraulic_conductivity_scale_parameter) = p.exponential
-    if z < p.z_exp[i]
-        kv_z =
-            vertical_hydraulic_conductivity_factor[i][n] *
-            kv_0[i] *
-            exp(-hydraulic_conductivity_scale_parameter[i] * z)
-    else
-        kv_z =
-            vertical_hydraulic_conductivity_factor[i][n] *
-            kv_0[i] *
-            exp(-hydraulic_conductivity_scale_parameter[i] * p.z_exp[i])
-    end
-    return kv_z
-end
-
-function hydraulic_conductivity_at_depth(
-    p::KvLayered,
-    vertical_hydraulic_conductivity_factor,
-    z,
-    i,
-    n,
-)
-    kv_z = vertical_hydraulic_conductivity_factor[i][n] * p.kv[i][n]
-    return kv_z
-end
-
-function hydraulic_conductivity_at_depth(
-    p::KvLayeredExponential,
-    vertical_hydraulic_conductivity_factor,
-    z,
-    i,
-    n,
-)
-    return if z < p.z_layered[i]
-        vertical_hydraulic_conductivity_factor[i][n] * p.kv[i][n]
-    else
-        n = p.nlayers_kv[i]
-        vertical_hydraulic_conductivity_factor[i][n] *
-        p.kv[i][n] *
-        exp(-p.hydraulic_conductivity_scale_parameter[i] * (z - p.z_layered[i]))
-    end
 end
 
 """
@@ -1060,21 +900,6 @@ function bounded_divide(x::Real, y::Real; max::Real = 1.0, default::Real = 0.0):
 end
 
 """
-    bounded_power(base, power)
-
-Computes min(base^power, 1) without computing the power
-if the result is known to be larger than 1.
-Assumes base, power > 0
-"""
-function bounded_power(base::T, power) where {T}
-    return if base > 1
-        one(T)
-    else
-        pow(base, power)
-    end
-end
-
-"""
 The sine of the slope in radians;
 sin(arctan(x)) = x / √(1 + x²)
 """
@@ -1127,9 +952,4 @@ function water_table_change(
     end
     exfilt = max(net_flux, 0.0)
     return dh, exfilt
-end
-
-"Set lower bound for drainable porosity"
-function lower_bound_drainable_porosity(theta_s, theta_fc; lower_bound = 0.02)
-    return max(theta_s - theta_fc, lower_bound)
 end
